+++ src/render/image_manager.ts
@@ -132,9 +132,9 @@ class ImageManager extends Evented {
         return true;
     }
 
-    updateImage(id: string, image: StyleImage) {
+    updateImage(id: string, image: StyleImage, validate = true) {
         const oldImage = this.images[id];
-        if (oldImage.data.width !== image.data.width || oldImage.data.height !== image.data.height) {
+        if (validate && (oldImage.data.width !== image.data.width || oldImage.data.height !== image.data.height)) {
             throw new Error(`size mismatch between old image (${oldImage.data.width}x${oldImage.data.height}) and new image (${image.data.width}x${image.data.height}).`);
         }
         image.version = oldImage.version + 1;